cosmetic updates, but still unabel to resolve VAE issue. going to merge back to develop and work on it later.
